requirements.txt
Django
djangorestframework
channels[daphne]
channels_redis
django-filter
